# note-keeper-web/swagger.yaml
openapi: 3.0.3
info:
  title: NoteKeeper API
  description: |
    Open-source Evernote alternative with powerful features for notes and todos management.
    
    ## Features
    - **Notes Management**: Create, read, update, delete notes with Markdown & Mermaid support
    - **Todos Management**: Manage todos with descriptions, tags, priorities, and schedules
    - **Folder Structure**: Organize notes in folders and subfolders
    - **Tags & Priorities**: Categorize and prioritize your content
    - **Encryption**: Secure sensitive notes
    - **Reminders**: Set reminders for notes and todos
    - **Archive & Trash**: Archive old items and restore from trash
    - **Templates**: Create notes from predefined templates
    - **History**: Track note changes with version history
    - **Search**: Full-text search with saved queries
    - **Calendar**: Visualize todos with deadlines
    - **Analytics**: Productivity graphs and statistics
    - **Shortcuts**: Keyboard shortcuts for quick actions
    - **Integrations**: Send notifications to Telegram, DingTalk, and Email
  version: 3.0.0
  contact:
    name: NoteKeeper Support
    email: support@notekeeper.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.notekeeper.example.com/v1
    description: Production server
...
paths:
  /notes:
    get:
      tags:
        - Notes
      summary: Get all notes
      description: Retrieve a list of all notes with optional filtering
      operationId: getNotes
      parameters:
        - name: folder
          in: query
          description: Filter by folder name
          schema:
            type: string
        - name: tag
          in: query
          description: Filter by tag
          schema:
            type: string
        - name: priority
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [low, medium, high]
        - name: isFavorite
          in: query
          description: Filter by favorite status
          schema:
            type: boolean
        - name: isEncrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
        - name: isArchived
          in: query
          description: Filter by archive status
          schema:
            type: boolean
        - name: isDeleted
          in: query
          description: Filter by deletion status
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Notes
      summary: Create a new note
      description: Create a new note with title, content, and metadata
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/{id}:
    get:
      tags:
        - Notes
      summary: Get note by ID
      description: Retrieve a specific note by its ID
      operationId: getNoteById
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Notes
      summary: Update note
      description: Update an existing note
      operationId: updateNote
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteInput'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Notes
      summary: Delete note (move to trash)
      description: Soft delete a note by moving it to trash
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
        - name: permanent
          in: query
          description: Permanently delete without moving to trash
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Note deleted successfully
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/{id}/archive:
    post:
      tags:
        - Notes
      summary: Archive note
      description: Move note to archive
      operationId: archiveNote
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
      responses:
        '200':
          description: Note archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/{id}/restore:
    post:
      tags:
        - Notes
      summary: Restore note
      description: Restore note from trash or archive
      operationId: restoreNote
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
      responses:
        '200':
          description: Note restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/{id}/history:
    get:
      tags:
        - Notes
      summary: Get note history
      description: Retrieve version history for a note
      operationId: getNoteHistory
      parameters:
        - name: id
          in: path
          required: true
          description: Note ID
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteHistory'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/import:
    post:
      tags:
        - Notes
      summary: Import note from file
      description: Import a note from MD, PDF, DOC, DOCX, XLS, or CSV file
      operationId: importNote
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to import (MD, PDF, DOC, DOCX, XLS, CSV)
                folder:
                  type: string
                  description: Target folder for imported note
                subfolder:
                  type: string
                  description: Target subfolder for imported note
      responses:
        '201':
          description: Note imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid file format or input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /search:
    get:
      tags:
        - Search
      summary: Search notes and todos
      description: Full-text search across notes and todos
      operationId: search
      parameters:
        - name: query
          in: query
          required: true
          description: Search query string
          schema:
            type: string
        - name: type
          in: query
          description: Filter by type
          schema:
            type: string
            enum: [all, notes, todos]
            default: all
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
        - name: priority
          in: query
          description: Filter by priority
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Note'
                  todos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /search/queries:
    get:
      tags:
        - Search
      summary: Get saved queries
      description: Retrieve all saved search queries
      operationId: getSavedQueries
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedQuery'

    post:
      tags:
        - Search
      summary: Save a search query
      description: Save a search query for later use
      operationId: saveQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedQueryInput'
      responses:
        '201':
          description: Query saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedQuery'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /search/queries/{id}:
    delete:
      tags:
        - Search
      summary: Delete saved query
      description: Delete a saved search query
      operationId: deleteSavedQuery
      parameters:
        - name: id
          in: path
          required: true
          description: Query ID
          schema:
            type: string
      responses:
        '204':
          description: Query deleted successfully
        '404':
          description: Query not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /integrations/telegram:
    post:
      tags:
        - Integrations
      summary: Send notification to Telegram
      description: Send a notification message to configured Telegram bot
      operationId: sendToTelegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Message to send
                  example: "New note created: Project Meeting Notes"
                botToken:
                  type: string
                  description: Telegram bot token (optional if configured in settings)
                chatId:
                  type: string
                  description: Telegram chat ID (optional if configured in settings)
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message sent to Telegram"
        '400':
          description: Invalid input or missing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /integrations/dingtalk:
    post:
      tags:
        - Integrations
      summary: Send notification to DingTalk
      description: Send a notification message to configured DingTalk webhook
      operationId: sendToDingTalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Message to send
                  example: "Reminder: Complete project documentation"
                webhook:
                  type: string
                  description: DingTalk webhook URL (optional if configured in settings)
                secret:
                  type: string
                  description: DingTalk secret (optional if configured in settings)
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message sent to DingTalk"
        '400':
          description: Invalid input or missing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /integrations/email:
    post:
      tags:
        - Integrations
      summary: Send email notification
      description: Send a notification email via configured SMTP
      operationId: sendEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subject
                - message
              properties:
                subject:
                  type: string
                  description: Email subject
                  example: "NoteKeeper Reminder"
                message:
                  type: string
                  description: Email message body
                  example: "You have 3 todos due today"
                to:
                  type: string
                  description: Recipient email (optional if configured in settings)
                  example: "user@example.com"
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Email sent successfully"
        '400':
          description: Invalid input or missing configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    NoteHistory:
      type: object
      required:
        - id
        - content
        - timestamp
        - action
      properties:
        id:
          type: string
          description: History entry ID
          example: "hist_1234567890"
        content:
          type: string
          description: Note content at this version
          example: "# Meeting Notes\n\nDiscussed project timeline."
        timestamp:
          type: string
          format: date-time
          description: When this version was created
          example: "2026-04-13T10:00:00Z"
        action:
          type: string
          enum: [created, edited, restored]
          description: Action type
          example: "edited"
    Note:
      type: object
      required:
        - id
        - title
        - content
        - folder
        - priority
        - isFavorite
        - isEncrypted
        - isArchived
        - isDeleted
        - attachments
        - history
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "1234567890"
        title:
          type: string
          description: Note title
          example: "Project Meeting Notes"
        content:
          type: string
          description: Note content (supports Markdown and Mermaid)
          example: "# Meeting Summary\n\n## Attendees\n- John Doe\n- Jane Smith\n\n## Action Items\n- [ ] Review proposal\n- [ ] Schedule follow-up"
        tags:
          type: array
          items:
            type: string
          description: List of tags
          example: ["work", "meeting", "important"]
        folder:
          type: string
          description: Folder name
          example: "work"
        subfolder:
          type: string
          description: Subfolder name (optional)
          example: "projects"
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: "high"
        isFavorite:
          type: boolean
          description: Favorite status
          example: true
        isEncrypted:
          type: boolean
          description: Encryption status
          example: false
        isArchived:
          type: boolean
          description: Archive status
          example: false
        isDeleted:
          type: boolean
          description: Deletion status (in trash)
          example: false
        deletedAt:
          type: string
          format: date-time
          description: Deletion timestamp
          example: "2026-04-13T12:00:00Z"
        reminder:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2026-04-15T10:00:00Z"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: List of file attachments
        history:
          type: array
          items:
            $ref: '#/components/schemas/NoteHistory'
          description: Version history
        templateId:
          type: string
          description: Template ID if created from template
          example: "tmpl_123"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-04-13T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-04-13T09:15:00Z"
    NoteInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: Note title
          example: "Project Meeting Notes"
        content:
          type: string
          description: Note content (supports Markdown and Mermaid)
          example: "# Meeting Summary\n\nDiscussed project timeline and deliverables."
        tags:
          type: array
          items:
            type: string
          description: List of tags
          example: ["work", "meeting"]
        folder:
          type: string
          description: Folder name
          example: "work"
          default: "default"
        subfolder:
          type: string
          description: Subfolder name (optional)
          example: "projects"
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: "medium"
          default: "medium"
        isFavorite:
          type: boolean
          description: Favorite status
          example: false
          default: false
        isEncrypted:
          type: boolean
          description: Encryption status
          example: false
          default: false
        reminder:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2026-04-15T10:00:00Z"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: List of file attachments
        templateId:
          type: string
          description: Template ID if creating from template
          example: "tmpl_123"
    Todo:
      type: object
      required:
        - id
        - title
        - description
        - completed
        - priority
        - isFavorite
        - isArchived
        - isDeleted
        - attachments
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "9876543210"
        title:
          type: string
          description: Todo title
          example: "Complete project documentation"
        description:
          type: string
          description: Todo description (supports Markdown)
          example: "Write comprehensive documentation for the API endpoints"
        completed:
          type: boolean
          description: Completion status
          example: false
        tags:
          type: array
          items:
            type: string
          description: List of tags
          example: ["documentation", "urgent"]
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: "high"
        isFavorite:
          type: boolean
          description: Favorite status
          example: true
        isArchived:
          type: boolean
          description: Archive status
          example: false
        isDeleted:
          type: boolean
          description: Deletion status (in trash)
          example: false
        deletedAt:
          type: string
          format: date-time
          description: Deletion timestamp
          example: "2026-04-13T12:00:00Z"
        dueDate:
          type: string
          format: date-time
          description: Due date
          example: "2026-04-20T17:00:00Z"
        reminder:
          type: string
          format: date-time
          description: Reminder date and time
          example: "2026-04-19T09:00:00Z"
        location:
          type: object
          properties:
            lat:
              type: number
              format: double
              example: 37.7749
            lng:
              type: number
              format: double
              example: -122.4194
            address:
              type: string
              example: "Office Building A, Room 301"
        schedule:
          type: object
          properties:
            repeat:
              type: string
              enum: [none, daily, weekly, monthly]
              example: "weekly"
            endDate:
              type: string
              format: date-time
              example: "2026-12-31T23:59:59Z"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: List of file attachments
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-04-13T08:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-04-13T09:15:00Z"
...
    SavedQuery:
      type: object
      required:
        - id
        - name
        - query
        - filters
        - createdAt
      properties:
        id:
          type: string
          description: Unique identifier
          example: "query_123"
        name:
          type: string
          description: Query name
          example: "Work todos"
        query:
          type: string
          description: Search query string
          example: "project documentation"
        filters:
          type: object
          properties:
            type:
              type: string
              enum: [all, notes, todos]
            tags:
              type: array
              items:
                type: string
            priority:
              type: string
              enum: [low, medium, high]
            folder:
              type: string
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-04-13T08:30:00Z"
    SavedQueryInput:
      type: object
      required:
        - name
        - query
      properties:
        name:
          type: string
          description: Query name
          example: "Work todos"
        query:
          type: string
          description: Search query string
          example: "project documentation"
        filters:
          type: object
          properties:
            type:
              type: string
              enum: [all, notes, todos]
            tags:
              type: array
              items:
                type: string
            priority:
              type: string
              enum: [low, medium, high]
            folder:
              type: string
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Title is required"
        details:
          type: object
          description: Additional error details
          additionalProperties: true